+++ NEWS.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `compute_cut` option to `EmbeddedCollection`. This indicates whether to compute cut
   geometries. If false, the recipes will be called with only `φh`.
 - Added `MultiLevelSetEvolution`
+- Added `rrule` for `restrict_to_field`
 
 ### Changed
 - Adjusted README.md to discuss AD and link to docs.